requirements.txt
# Data Processing & Analytics
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Azure Cloud Services
azure-storage-file-datalake>=12.14.0
azure-storage-blob>=12.19.0

# Configuration
python-dotenv>=1.0.0

# HTTP / API
requests>=2.31.0

# Email Templating
Jinja2>=3.1.0

# Excel Support (pandas engine)
openpyxl>=3.1.0
